SCHED-298: Build parent component
