+++ src/components/ControlPanel/styles.scss
@@ -40,3 +40,49 @@
         box-shadow: 0 0 0 1px #499bdb;
     }
 }
+
+
+.stbtn-container {
+    padding-inline: 10px 10px;
+    margin-right: 0;
+
+    .p-selectbutton{
+        .p-buttonset{
+            padding-top: 2px;
+            padding-left: 0;
+        }
+        .p-button {
+        margin-right: 0;
+        padding-top: 3%;
+        padding-bottom: 3%;
+
+        }
+
+    }
+
+    .p-component.p-button {
+        margin-right: 0 !important;
+    }
+}
+
+.toggle-btn {
+    background: none;
+    border: 0px;
+    font-size: 13px;
+    margin-right: 0;
+    padding: 0;
+    .p-component{
+        .p-button{
+            margin-right: 0;
+        }
+    }
+    
+}
+
+toggle-btn:hover {
+    background: none ; 
+}
+
+
+  
+
