--- src/components/DatePicker/DatePicker.scss
@@ -1,2 +1,0 @@
-@import 'lucuma-ui-css/lucuma-ui';
-@import 'lucuma-ui-css/dark-theme';
