--- src/components/DatePicker/DatePicker.tsx
@@ -1,23 +1,0 @@
-import { useState } from 'react';
-import { Calendar } from 'primereact/calendar';
-import './DatePicker.scss';
-
-
-export default function DatePicker(){
-
-    const [dates2, setDates2] = useState<Date | Date[] | string | null | undefined>(undefined);
-   
-
-    return(<>
-            
-            <Calendar id="range" 
-                value={dates2} 
-                onChange={(e) => setDates2(e.value)} 
-                selectionMode="range" 
-                readOnlyInput 
-                showButtonBar
-                showIcon/>
-       
-            </>
-    );
-}
