--- src/components/SiteToggleButton/SiteToggleButton.scss
@@ -1,45 +1,0 @@
-
-.stbtn-container {
-    padding-inline: 10px 10px;
-    margin-right: 0;
-}
-
-.toggle-btn {
-    background: none;
-    border: 0px;
-    font-size: 13px;
-    margin-right: 0;
-    padding: 0;
-    .p-component{
-        .p-button{
-            margin-right: 0;
-        }
-    }
-    
-}
-
-toggle-btn:hover {
-    background: none ; 
-}
-
-
-.p-selectbutton{
-    margin: 0 0 0 0; d
-    .p-buttonset{
-        padding-top: 2px;
-        padding-left: 0;
-    }
-    .p-button {
-       margin-right: 0;
-       padding-top: 3%;
-       padding-bottom: 3%;
-
-    }
-
-}
-
-.p-component.p-button {
-    margin-right: 0 !important;
-}
-  
-
