--- src/stories/DatePicker.stories.tsx
@@ -1,14 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-
-import DatePicker from '../components/DatePicker/DatePicker';
-
-export default {
-    title: 'Scheduler/Date picker',
-    component: DatePicker,
-} as ComponentMeta <typeof DatePicker>;
-
-const Template: ComponentStory<typeof DatePicker> = () => <DatePicker/>;
-
-
-
-export const Normal = Template.bind({});
